P1-Gym Member Workout & Progress Tracker system/backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-jose[cryptography]
python-multipart
passlib[bcrypt]
pydantic
streamlit
pandas
matplotlib
pytest
httpx
